+++ arch/cpu/cc2538/rtimer-arch.h
@@ -66,8 +66,8 @@
 /* Do the math in 32bits to save precision.
  * Round to nearest integer rather than truncate. */
 #define US_TO_RTIMERTICKS(US)  ((US) >= 0 ?                        \
-                               (((int32_t)(US) * (RTIMER_ARCH_SECOND) + 500000) / 1000000L) :      \
-                               ((int32_t)(US) * (RTIMER_ARCH_SECOND) - 500000) / 1000000L)
+                               ((int32_t)(((int64_t)(US) * (RTIMER_ARCH_SECOND) + 500000) / 1000000L)) :      \
+                               ((int32_t)((int64_t)(US) * (RTIMER_ARCH_SECOND) - 500000) / 1000000L))
 
 #define RTIMERTICKS_TO_US(T)   ((T) >= 0 ?                     \
                                (((int32_t)(T) * 1000000L + ((RTIMER_ARCH_SECOND) / 2)) / (RTIMER_ARCH_SECOND)) : \
